+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -1,4 +1,6 @@
 type IconName =
+	| "aiChatActive"
+	| "aiChatInactive"
 	| "check"
 	| "clockActive"
 	| "clockInactive"