+++ apps/frontend/src/modules/categories/libs/types/types.ts
@@ -1,1 +1,5 @@
-export { type CategoriesGetAllResponseDto, type CategoryDto } from "shared";
+export {
+	type CategoriesGetAllResponseDto,
+	type CategoriesGetRequestQueryDto,
+	type CategoryDto,
+} from "shared";