+++ apps/mobile/src/libs/helpers/get-gradient-colors-for-category/get-gradient-colors-for-category.helper.ts
@@ -1,0 +1,14 @@
+import { categoryToColors, type colorToGradientColors } from "~/libs/maps/maps";
+
+const DEFAULT_SLIDER_COLOR = "blue";
+
+const getGradientColorsForCategory = (
+	categoryName: string,
+): keyof typeof colorToGradientColors => {
+	return (
+		categoryToColors[categoryName.replaceAll(/\s+/g, "")] ??
+		DEFAULT_SLIDER_COLOR
+	);
+};
+
+export { getGradientColorsForCategory };