+++ apps/mobile/src/slices/quiz/quiz.slice.ts
@@ -8,7 +8,7 @@ import {
 	type QuizScoresGetAllItemResponseDto,
 } from "~/packages/quiz/quiz";
 
-import { getAllQuestions, getScores } from "./actions";
+import { editScores, getAllQuestions, getScores } from "./actions";
 
 type Properties = {
 	answerId: number;
@@ -56,6 +56,31 @@ const { actions, name, reducer } = createSlice({
 		});
 		builder.addCase(getAllQuestions.rejected, (state) => {
 			state.dataStatus = DataStatus.REJECTED;
+		});
+		builder.addCase(editScores.fulfilled, (state, action) => {
+			const updatedScores = new Map(
+				action.payload.items.map((score) => [score.id, score]),
+			);
+
+			state.scores = state.scores.map((stateScore) => {
+				const updatedScore = updatedScores.get(stateScore.id);
+
+				return updatedScore
+					? {
+							...stateScore,
+							score: updatedScore.score,
+							updatedAt: updatedScore.updatedAt,
+						}
+					: stateScore;
+			});
+
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(editScores.rejected, (state) => {
+			state.dataStatus = DataStatus.REJECTED;
+		});
+		builder.addCase(editScores.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
 		});
 	},
 	initialState,