+++ packages/shared/src/modules/categories/categories.ts
@@ -1,10 +1,12 @@
 export { CategoriesApiPath } from "./libs/enums/enums.js";
 export {
 	type CategoriesGetAllResponseDto,
+	type CategoriesGetRequestQueryDto,
 	type CategoriesSelectedRequestDto,
 	type CategoryCreateRequestDto,
 	type CategoryDto,
 	type CategoryUpdateRequestDto,
 	type CategoryWithScoresDto,
 } from "./libs/types/types.js";
+export { categoryIds as categoryIdsValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
 export { categoriesSelected as categoriesSelectedValidationSchema } from "./libs/validation-schemas/validation-schemas.js";