+++ packages/shared/src/modules/categories/libs/validation-schemas/category-ids.validation-schema.ts
@@ -1,0 +1,19 @@
+import { z } from "zod";
+
+import {
+	CategoriesValidationMessage,
+	CategoriesValidationRegexRule,
+} from "../enums/enums.js";
+
+const categoryIds = z.object({
+	categoryIds: z
+		.string()
+		.trim()
+		.regex(
+			CategoriesValidationRegexRule.QUERY_IS_STRINGIFIED_INTEGERS_ARRAY,
+			CategoriesValidationMessage.NOT_INTEGER_ARRAY,
+		)
+		.optional(),
+});
+
+export { categoryIds };